Relocated booleans and numbers to the common directorY
